currency-converter/requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
